requirements.txt
streamlit
pytesseract
PyPDF2
Pillow
requests